fix: rename kubectl extension
I'll be honest and admit that I haven't spent the time to figure out exactly
_why_ this works, but I assume it is due to the build script or elsewhere
assuming that the extension name matches the folder name. Either way, it is
a better/shorter name for this extension, suggested by Florent during the
initial review, and causes the extension to correctly be included in the
production build.

Signed-off-by: Tim deBoer <git@tdeboer.ca>