fix: don't refresh image list when age updates
We have been replacing the entire image list whenever the age is updated.
This worked (?) for the <table> implementation, but for the new table
component it is triggering a full refresh/reset of the table. Just updating
the image age properties and then setting images to itself is enough to
trigger Svelte to update the UI without a full replacement.

Signed-off-by: Tim deBoer <git@tdeboer.ca>