feat: form progress (#5253)
* feat: form progress

I'm pretty sure this was an idea from Mo. Adds a linear progress bar at the
bottom of the FormPage header to indicate that something is 'in progress',
and uses this when building and pulling images, playing kube, deploying pods,
creating pods from containers, creating volumes, and creating resources
(container or Kubernetes).

This helps to indicate that there's 'something happening', especially in
cases like #4229 where you can't see another progress monitor in the view.

This 'duplicates' the progress monitor when creating a container/Kubernetes
instance, but I don't want to remove that as these pages are also used during
onboarding and the original/smaller one helps draw your eye to the Show Logs
expander. IMHO having two here is still an improvement.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: add tests

When you write a test, it helps if you include it in the commit. :)

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>