chore: switch provider cards to icon
The updated dashboard design in #4395 uses provider icons (square) and not
logos (could be wider or taller). This changes switches from using .logo to
.icon, and from using ProviderLogo to IconImage, since we don't need
ProviderLogo anymore and IconImage also supports light mode. Tests added.

I would like to delete ProviderLogo but it is still used by
PreferencesProviderInstallationModel - so that will be a different PR once
I have a chance to do the same change there.

Signed-off-by: Tim deBoer <git@tdeboer.ca>