feat: duration column
Creates a new duration column that takes a date and displays a
humanized time, then automatically refreshes exactly when there
should be an update.

Since the timeouts and UI are handled by the column itself there's no
need for the UI object to store and update the age string, for the
table to have timers or do anything related to the update.

Core fix for #4972, would still need to be used by other tables.

Signed-off-by: Tim deBoer <git@tdeboer.ca>