fix: empty screen reset filter by default
Fixes the regression caused by #4600. Prior to this the FilteredEmptyScreen
would reset all filters to ''. The container list needed special processing
to not remove the filters like 'is:running' so it was changed to an event,
but none of the other lists respond to this event so none of the filters get
cleared.

This sets a default behaviour of clearing the list so that none of the other
lists need to respond to the event if they are fine just clearing the filter.
If you do your own special processing, you just call e.preventDefault(). The
search term is included as an event property because we need to set
cancelable in the third argument.

The pods list has the tabs/is:running so I added the event handling there as
well. Without this change clearing the filter would always return you to the
first tab.

Fixes #5288.

Signed-off-by: Tim deBoer <git@tdeboer.ca>