fix: stop provider card overflow
Small tweaks to ProviderCard to make sure the icon margin isn't growing and
right column is growing correctly. Uses max-w-xs to limit the width of the
right column wherever text was not wrapping correctly. Removes redundant
provider version from the configured state.

Even with these changes it is still possible to get overlap when the
window is between 640px (the minimum) and 690px wide. We're at the point
where the required content does not fit the current design. If anyone else
has a suggestion I'm willing to try it, but IMHO it may not be worth it as
this is more of an extreme/edge case now, and we've been waiting for a
couple sprints now on a design for moving some of the actions to other
places like Resources.

Since this is visual overlap of elements not staying within their flex
layout I can't think of any useful tests.

Fixes #5316.

Signed-off-by: Tim deBoer <git@tdeboer.ca>