feat: table update event passthrough
The pods and containers list use an update event (previously callbacks)
to let the parent page know that one column has triggered a change that
requires a UI refresh. As we move to the table component there is an
additional level of components, so we just need this event to pass through
the table component.

This change just allows the table to pass through update events. To test
the age column was changed to trigger a callback whenever it is sorted.

Signed-off-by: Tim deBoer <git@tdeboer.ca>