feat: Deployments & Service details pages
This is Details pages for both Deployments and Services. Name columns
are updated to buttons that link to details pages, just like in
other pages.

Compared to what we have for pods I did two things differently:
- kube objects are read once for all tabs, instead of each tab loading
  them independently.
- Due to this we can make Inspect and Kube tabs simple Monaco components
  (basically just JSON and YAML versions of the same thing) instead of
  having additional Svelte components.

For the Summary page I added a few fields to start, and tried reusing some
of the table columns. We will probably want more information, but this
start with a basic set.

Fixes #5157.
Fixes #5159.

Signed-off-by: Tim deBoer <git@tdeboer.ca>