feat: image status & deletion
Images had an inUse flag that we used to determine which status icon
to show or whether the image is delete-able. If you delete the image
and there is any delay, there is no feedback in the UI and you can
delete it again (or accidentally click on the next image delete if the
first one disappears at the wrong time).

This changes images to have a status the same as most other types (e.g.
pods, containers, deployments), so once the user clicks to delete it is
marked for deletion (spinner in UI) and you can't delete again.

Fixes #5474.

Signed-off-by: Tim deBoer <git@tdeboer.ca>