feat: make Kubernetes support experimental
Adds a new experimental flag (off by default) that hides the Kubernetes
options on the left navbar. When you enable it, the options automatically
appear.

Tests added to make sure that the Kubernetes options don't appear if the
preference is not set, and do appear if it is (and have at least one
context).

Fixes #5525.

Signed-off-by: Tim deBoer <git@tdeboer.ca>