feat: merge extension preferences
Sets the same id and title on the extension update preferences as the
activation pref. We could merge the two, but having two contributions
with the same id doesn't seem to have any side effects and this keeps
each bit of code separate.

Yes, this means that anyone who has set these two properties will have
them reset when they update to this PR, but that seems inevitable when
we fix this (the sooner the better) and doesn't seem worth additional
code to automatically migrate the setting.

Fixes #5549.

Signed-off-by: Tim deBoer <git@tdeboer.ca>