fix: deployment conditions overflow (#5531)
* fix: deployment conditions overflow

While the deployments PR was up for review, we changed the default for
columns to not allow overflow unless the column specifically requests it.
This makes horizontal scaling avoid overlapping columns, but columns like
conditions need to specifically request overflow for tooltips. This just
adds the missing attribute.

Fixes #5463.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: add test

Adds test to confirm that conditions is the only column that has overflow
enabled. (we'll want to add it to actions too sometime, but unnecessary
for now, and we have the test to prove when we add it).

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: lint

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>