chore: kubeconfig topic
After doing other PRs I noticed that the Kubeconfig preference has an odd path.
This changes it to just 'Kubernetes' to fit in and be common enough to sit
with other Kubernetes options.

This is a 'breaking change' since it will reset the preference for anyone who
has changed it, but I don't think it is worth trying to migrate it and doing
it now (before we start to release more Kubernetes UI support) is much better
than holding off.

Fixes #5551.

Signed-off-by: Tim deBoer <git@tdeboer.ca>