feat: ingress & route details pages
Basic details pages for ingress and routes. As with #5413 we'll need
to add more details to the summary, but at least this provides the
basic support so that you can click on an ingress or route and see
the basic summary, Inspect, and view YAML.

I've gone with two separate detail pages instead of a combined page.
Although the list was a merged page, this makes it easier for each
page to follow the basic pattern (e.g. automatically closing the
page if you're looking at something that is deleted) and have a
title that says 'Ingress Details' instead of 'Ingress or Route
Details'. The odd choice of '/ingressesRoutes/ingress' routes are so
that the main nav keeps the ingress/route icon selected when you
go to details, just like the other pages.

Signed-off-by: Tim deBoer <git@tdeboer.ca>