fix: quick pick filter removes selection
As you filter the quick pick would automatically select the first item
that's left, but would never deselect, so (e.g.) if you type 'ddd' you're
selecting whatever the last item that started with 'd' was (or the first
item ever shown if there were none that started with d). This just adds the
deselection if there are no items that match the filter.

I decided to use the 'onSelectCallback' to test and found a second problem:
the callback is not called when the selection is changed due to filtering.
That is fixed and tested as well.

Fixes #5610.

Signed-off-by: Tim deBoer <git@tdeboer.ca>