chore: image status
Images have an inUse flag that we use to determine which status icon
to show or whether the image is deletable. However, that means we
cannot have other states like deleting or new.

This changes images to have a status the same as most other types (e.g.
pods, containers, deployments). This will allow future PRs to add
additional states (e.g. deleting) and UI updates, e.g. so that you see
a spinner if an image takes a second to delete.

Step toward #5474.

Signed-off-by: Tim deBoer <git@tdeboer.ca>