chore: change style of unknown OS download button
When we can determine the user's OS, we have a nice purple Download
Now button on the front page.
When we can't determine the user's OS, it's a full-width blue
button linking to the Download Page instead. This button doesn't
match our current styling at all.

I copied from the Download Now button to make it match:
- Added outer div (makes it normal size and centered).
- Changed indigo to purple.
- Added 'mt-6 mb-1' to space it from above and below.

Signed-off-by: Tim deBoer <git@tdeboer.ca>