chore: fix color of server-side download button
The download button on the front page of the website is generated
client-side, and has a fallback in case we can't detect the user's
browser.

Before any of this happens, docusaurus generates the button on the
server-side, which is essentially a fallback that should never be
seen. This button is the last thing in our codebase using indigo
(blue) and is missing the mt-6 mb-1 spacing from the client-side
version.

This change just fixes the color and mt/mb so that the button is
identical to the client-side one on line 72. (matches current
style)

Signed-off-by: Tim deBoer <git@tdeboer.ca>