chore: use status colors
There are status colors in the palette, but prior code is referring
to the color values directly. This is just cleanup to make everything
use the status color.

Adds one 'missing' status color for starting, and there is one slight
color change (degraded will go from amber-600 to the now official
amber-700), but this is not a noticeable change. I did a scan for
other references and didn't find any.

Fixes #4740.

Signed-off-by: Tim deBoer <git@tdeboer.ca>