feat: input component
Creates a new Input component with the same basic support as an
HTML input, but with our styling. Supports 4 types so far:
- normal text
- search (adds a search icon on the left)
- clear (adds an X on the right, which clears the field unless
  the parent event handler overrides it)
- password (with a show/hide button on the right ( unless the
  parent event handler disables it)

I don't want to give any impression that this component is
fully done or supports all cases, my intention is to get a good
base component that does enough to support the initial use, then
expand it as necessary to support different uses or additional
style requirements.

In order to give something to test, the first few uses of Input
are included:
- Main nav page search bar
- Settings > Preferences search bar
- Settings > Preferences string and file items

First step of #3234.

Signed-off-by: Tim deBoer <git@tdeboer.ca>