feat: kube apply yaml button
Creates an 'Apply YAML' button that prompts for a file, applies it to
the current context, and reports how many resources have been created.

I've added the button to the Deployments, Services, and Ingress/Routes
pages, as these are the most useful pages for a Kubernetes user. I
didn't add it to the Pods page b/c this is less likely used for a
Kube user and it might create confusion with the 'Play Kubernetes YAML'
button.

Created a SolidKubeIcon identical to how we created a SolidPodIcon
for the 'create pod from containers' action.

Signed-off-by: Tim deBoer <git@tdeboer.ca>