chore: extract password/search components
Based on PR feedback - instead of a type this leaves Input as a basic/
general purpose input and exposes much simpler components SearchInput
and PasswordInput that extend it with the expected icons and actions.

I had to make some choices of what to expose from Search/PasswordInput
and tried to keep them fairly minimal. SearchInput has a clear icon and
I updated the ContainerList test b/c I went with a more standard '...'.

Signed-off-by: Tim deBoer <git@tdeboer.ca>