nit: status bar should have z-40 to go above onboarding button area
### What does this PR do?

Small nitpick, there is an overlap from the onboarding buttons which are
z-20 to z-30.

The status bar should be z-40 so it'll always appear above the buttons /
not overlap.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

<!-- Please explain steps to reproduce -->

Go to onboarding sequence and see that there is no more "overlapping".

Signed-off-by: Charlie Drage <charlie@charliedrage.com>