nit: make guide buttons smaller
### What does this PR do?

I find that the buttons are massive compared to the other parts of the
dashboard.

PR to make the buttons a tad smaller (medium) which matches the rest of
the dashboard.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

N/A

### How to test this PR?

`yarn watch` and look at the dashboard.

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>