chore: input errors, use input component in run image
Uses the Input component in the Run Image form. This is the first
time that we have inputs with 'errors', so I've added support to
Input for an error. When there's an error the purple changes to
red, there's a red X on the right, and the error message is listed
underneath the input.

Another part of #3234.

Signed-off-by: Tim deBoer <git@tdeboer.ca>