feat: input component in play kube
Use the Input component when playing a Kube YAML.

When I went to test this I was so confused how to select a file I
had to look at the code, and realized you're supposed to click on
the readonly text input.

This wasn't a great pattern before, but it's less obvious now that
we have disabled styling on Input, so I switched this to the more
common pattern of having a Browse button to open the file dialog.

Signed-off-by: Tim deBoer <git@tdeboer.ca>