chore: use checkbox component
I noticed several places where we're using a standard blue <input
type="checkbox" /> instead of our styled Checkbox. Run Image is the
biggest user.

This change originally caused failures in the E2E tests because our
Checkbox was incompatible with Playwright, since the input in this
checkbox component was neither visible nor Aria compliant.

I wasted lots of time earlier this year trying to build a custom
checkbox that was aria-compliant, but the fix was much simpler after
carefully reading the Playwright requirements here:
https://playwright.dev/docs/actionability#visible

The input checkbox cannot be hidden or 0 size, and cannot be 'behind'
the div icon. However, it can be opacity-0 and on top using absolute
positioning and changing the z-order by swapping the order of elements.

Signed-off-by: Tim deBoer <git@tdeboer.ca>