chore: use checkbox component
We are using standard checkboxes in the Deploy to Kubernetes
form instead of our styled Checkbox component. This just switches
them and updates the test to be more consistent.

Wrapper <div>s added because Checkbox (a composite) doesn't work
as well w/o a layout, but this also fixes the problem where text
was wrapping below the checkbox when it shouldn't.

Signed-off-by: Tim deBoer <git@tdeboer.ca>