feat: file input
We have Inputs next to Browse buttons in many places, all with slight
variations in UI or browsing. This creates a new FileInput component that
uses a regular Input component with a folder icon on the right to browse
for files. Exposes OpenDialogOptions to make it easy to customize what
the file dialog should filter for.

Uses the new FileInput for file/folder preferences so that there's something
to test. There are several other file/browse buttons elsewhere, but they
would be migrated separately. Likely also a candidate for svelte-ui package,
but that should be after we become comfortable with/use it.

Fixes #7197.

Signed-off-by: Tim deBoer <git@tdeboer.ca>