feat: add Red Hat Container Registry to default registries
The Red Hat Container Registry (`registry.redhat.io`) is widely used in
the ecosystem.  To make it easier for users to access and explore the
registry, add it to the list of default registries.  Since the
registries of other companies (e.g., Docker, Google, GitHub) are listed
as well, I do not perceive this change as forcing Red Hat products onto
users but as an UX improvement.

The main motivating driver behind this change was the immediate need of
the subscription/sso extension: showing the registry without the logo
is undesireable.  However, while re-reading throught the conversation I
felt that we may just add the registry by default and make it easier for
all users.

[1] https://github.com/redhat-developer/podman-desktop-redhat-account-ext/issues/38

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>