chore: use components in quickpick
The quickpick was using an unstyled checkbox and button, this just
switches it to use the Checkbox and Button components.

Checkbox required the ability to set the class. I also replaced two
violets with purple to match our normal selection color - although
it'll probably be replaced by variables soon, this will make it
match in the meantime and easier to find.

Signed-off-by: Tim deBoer <git@tdeboer.ca>