feat: consistent close button
We have 9 close buttons that I could find in Podman Desktop. They
were at least 3 different colors, 6 of them changed color when you
hover, and one hover changed the hover background too. Except for
the one, the lack of/inconsistent hover indication was standing out
to me compared with our Links and Buttons.

This provides a common CloseButton component in a medium-dark gray.
On hover, it goes white and has the same bg as a Link (but slightly
larger since the icon is small). The button accepts either an href
or an on:click action to support our different uses.

I've made a call here that all close buttons look the same, instead
of leaving those on dialogs/modals white. IMHO this consistency is
better, allows the hover indication to be more clear, and we don't
need to highlight the Close button in any of these places.

Although we might normally divide up the component + usage in two
PRs, the amount of change here wasn't big and I thought it's
important to see it in all of the contexts, so I've included both
and can split up if requested.

Fixes #2352.

Signed-off-by: Tim deBoer <git@tdeboer.ca>