fix: center empty screens
The pull/build image forms do not center the 'empty screen'. The
difference between them and other forms is h-fit vs h-full, changing
this fixes the issue and doesn't cause any regression when not
showing the empty screen.

Create volume has the same issue - although it's not available
when there is no container engine I thought we should keep the
forms consistent.

Fixes #6059.

Signed-off-by: Tim deBoer <git@tdeboer.ca>