draft: full kube informers integrated
After seeing the filtered stores in #6101 I thought I'd see how hard it
would be to do the same for ingress/routes/services. This is a quick
pass at:
- Adding informers for the three missing types to
  main/kubernetes-context-state.ts. This is mostly copy/paste from
  kubernetes-client.ts, only missing one issue on line 331).
- Adding filtered stores to renderer/kubernetes-contexts-state.ts.
- Switching all the pages over to use these stores.

These changes would clearly have to be broken up into multiple PRs and
tests updated, but we're not affecting existing UI until #3. After this,
the other stores could be removed, as well as informers in
kubernetes-client.ts.

Compared to the existing informers, this would fix the existing issues
in #5311 with initial connection/reconnection, and would add backoff,
but there are two issues we'd have to resolve:
- #6114 - reduce backoff for the current context, it's a bit slow.
- #6094 - we're probably creating too many informers; only loading
  these new ones for the current context would be enough for now.

Fixes #5311. Fixes #6113.

Signed-off-by: Tim deBoer <git@tdeboer.ca>