chore: number input
The current number editor in Settings > Preferences works, but it has a
few drawbacks:
 - doesn't match current design, and would need updating to light mode.
 - periodically has some odd 'resets' to a previous number. (this isn't
   just the input - figured out later this is #7142)
 - has two error modes - underline for below minimum, icon to the left
   for invalid value.
 - after invalid entry (e.g. below minimum), incrementing does a string
   add vs numerical.
 - not easily reusable for editing numbers in other places.

This creates a NumberInput component that is just an Input with +/- buttons.
Key input is restricted to numbers, and increment/decrement are used to
disable +/- appropriately. If the user manually enters a number out of
bounds, the Input's error is used to display a short warning.

Input required some minor changes to support this:
 - allow inputClass in order to center numbers in the textbox.
 - passthrough of key events, in order to block non-numerical entry.
 - boolean to block showing the error in a separate span. Although this is
   the design for inputs, it felt really out of place in preferences.
 - reduce padding slightly - it didn't matter in large Inputs, but
   number inputs tend to be smaller and extraneous padding was obvious.

New component is used in Settings > Preferences and max retry count for
running an image.

Tests added. NumberItem test had to be changed since only numerical inputs
work now.

Another part of #3234.

Signed-off-by: Tim deBoer <git@tdeboer.ca>