chore: switch buttons in run image to Button component
Switches all the buttons in the Run Images form to the Button
component. The normal button type is way too bold and stands out, and
the secondary button type has an outline and also stands out a lot,
so I've gone with 'link' since that matches the original look the most.

The Input component was using w-full where it should have been using
grow, which was causing it to overlap with buttons. Fixed and updated
test.

Fixes one part of #6220.

Signed-off-by: Tim deBoer <git@tdeboer.ca>