chore: switch ingress & routes to new informer
Provides ingress and routes informers using the same pattern as the
deployment and services informers, then switches the corresponding page
to use these too.

The pattern is identical to existing code, with the exception of the
context-state tests where I used describe.each() to iterate over all
secondary informers instead of maintaining separate tests for each one.
backoffInitialValue was raised to 10000 as it was hitting the limit
with multiple informers.

Signed-off-by: Tim deBoer <git@tdeboer.ca>