chore: remove previous informers from renderer
We've switched to using the newer per-context Kubernetes informers and are
not using the original informers anymore. As proof of that, to clean up the
codebase, and to stop us from accidentally using them, this removes all of
the informer stores from the renderer.

After this, we could remove the corresponding api bridge and the informers
themselves from main.

Signed-off-by: Tim deBoer <git@tdeboer.ca>