chore: switch details pages to new informers
Switches our 4 Kubernetes details pages over to the newer informer stores.
The changes to the pages themselves are very simple, but the old stores
were (likely incorrectly) writable and the new stores aren't, so mocking
was added to the tests to allow us to update the stores directly.

Fixes another part of #6259.

Signed-off-by: Tim deBoer <git@tdeboer.ca>