chore: add podman is installed content screen for podman onboarding
### What does this PR do?

Adds "Podman is installed" onboarding content similar to how we have it
for both Compose and Kubectl.

This gives the user some more information on what to do with the
installed podman CLI.

This also provides more information in case the onboarding is "re-ran"
again.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes https://github.com/containers/podman-desktop/issues/6269

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

1. Go through the onboarding again
2. See the final screen has been updated

Signed-off-by: Charlie Drage <charlie@charliedrage.com>