chore: align registry actions right
When we merged #5965 there was a comment that some of the actions used to
be right-aligned (it wasn't totally consistent) but our design language
always has actions in the last 'column' aligned to the right. I said I'd
follow up, this just changes all of the actions to be right for consistency.

Signed-off-by: Tim deBoer <git@tdeboer.ca>