nit: no right margin in menus
Our drop-down menus have a right margin, which means there is always a small
unused/white-space to the right.

The margin was added in Nov 2022 when the hover behaviour was added and I
can't find any reason for it. (maybe the main pages at the time needed the
actions moved from the right edge of the window?) I've been through
everywhere we use this component today, and removing the margin had the
same minor layout improvement in all cases.

Signed-off-by: Tim deBoer <git@tdeboer.ca>