fix: consistent button height
The height of primary/secondary/link buttons is inconsistent, because some
have borders and some don't. Secondary buttons get smaller when they're
in-progress or disabled, because the border is removed.

This modifies the default y padding to account for borders, and change
the secondary button to always have a border.

I tried making all the buttons have border (to match the current bg where
applicable) but that was more change, changed the styling slightly, and I
couldn't find a good way to match border opacity for the link button.
The multiple ? is ugly, but far shorter than that alternative.

Tab height unchanged (31h).

Before:
- link/primary/disabled secondary: 29.5h
- secondary: 31.5h (29.5 when disabled)
- danger: 33.5h

After: all 29.5h

Fixes #6304.

Signed-off-by: Tim deBoer <git@tdeboer.ca>