chore: dropdown component
Creates a new Dropdown component that is just a styled select control,
so that we don't have slight variations and styling in several places,
and one less thing to fix for light-mode. Styling uses the same color
pattern as an Input for consistency.

I used the new component in three places to start so that there's
something to test: deploying a pod to Kube, creating a pod from
containers, and enum items in preferences. There are 11 other instances
of <select> that will need to be switched in future PRs.

Signed-off-by: Tim deBoer <git@tdeboer.ca>