chore: improve incremental progressbar
Fixed several issues in the progress bar:

No progress is shown for tasks that have incremental (i.e. 5%, 10%) progress.
- Width was missing a ':'.

There is no visual animation for incremental tasks.
- Added a slight opacity pulsing animation.

The color for incremental vs indeterminate are different, and there's a hard-coded
rgb value.
- Indeterminate was set to purple-500 but then animation was hardcoded to use the
  purple-600 rgb value. Changed both to use purple-600 and no more hardcoded rgb.

The width and height of the bar is fixed and there's no layout, which would make
it harder to reuse.
- Moved the text inside the outer div and use flex layout. Turned width and height
  into parameters (and made slightly wider by default, since we have the room).

Fixes #6531.

Signed-off-by: Tim deBoer <git@tdeboer.ca>