feat: support for webview reveal (#6546)
* feat: support for webview reveal

When you call a Webview panel's reveal() function today, it fires a
webview-panel-reveal event that nothing listens to, so the panel is never
opened. There are a few possible ways we could solve this:
- Webview panel could call the navigation manager directly.
- Webview panel could fire a navigation event itself.
- Renderer can listen to the reveal event and navigate.

The first requires the navigation manager to be passed into the Webview, which
involves a decent amount of change and a somewhat circular dependency because
the navigation manager already depends on the webview registry. The third is
really simple and works, but this is essentially a second, custom webview
navigation event.

So this is the middle option: just trigger a webview navigation event instead.
Although it's not necessary the code uses NavigationPage and casts to
NavigationRequest to ensure type safety, and being caught if there are ever
changes to the navigation code.

The optional preserveFocus is (still) not respected, since webview
navigation doesn't support an option like this; we could add that separately,
whether or not we expose it via the navigation API.

Fixes ##5925.

Signed-off-by: Tim deBoer <git@tdeboer.ca>