chore: add isManifest function to determine a manifest
### What does this PR do?

Adds a guessIsManifest function to try and determine if an image is a
manifest or not:

- By checking the repotags and digests
- Checking labels does not exist, as labels are not added when creating
  a manifest
- VirtualSize is under 1MB as a manifest list is just text that points
  to other images

We are unable to check the history (manifest commonly has no history) as
Dockerode API does not support getting the history of the image.

### Screenshot / video of UI

<!-- If this PR is changing UI, please include
screenshots or screencasts showing the difference -->

N/A

### What issues does this PR fix or reference?

<!-- Include any related issues from Podman Desktop
repository (or from another issue tracker). -->

Closes  containers#6690

### How to test this PR?

<!-- Please explain steps to verify the functionality,
do not forget to provide unit/component tests -->

- [X] Tests are covering the bug fix or the new feature

Signed-off-by: Charlie Drage <charlie@charliedrage.com>