chore: include extension version in telemetry
The loadExtension and activateExtension telemetry events include the ids of
extensions but not the versions. It would be handy if we have them here both
for general debugging/fleet understanding, but also to stop each new
extension from wanting to add a startup event.

Extensions should not get to either path without a valid manifest + version,
just being slightly paranoid and consistent with prior code using ?.

There was no basic loadExtension test so I started one; also updated the
activateExtension test to check for telemetry.

Signed-off-by: Tim deBoer <git@tdeboer.ca>