chore: align DD extension actions
The action buttons for DD don't align with other extensions. This just adds a
dummy Stop button that is never enabled so that the buttons visually match the
other extensions.

There are other ways to solve this (changing the row layout, or adding a spacer
that matches the width of the action buttons) but IMHO this keeps DD extensions
as visually similar and consistent as possible to other extensions. From a user
perspective I don't think this is a negative b/c it's clear 'this can't be
stopped' vs potentially wondering 'why is there no stop button' before.

One part of #6855.

Signed-off-by: Tim deBoer <git@tdeboer.ca>