chore: remove kube list telemetry
kubernetesListNamespacePod is the single most common event sent in telemetry by
far, and responsible for a significant portion of our events. It doesn't really
tell us anything (we had access to pods on a namespace) compared with other
events like reading individual items or modifying them, and even logging the
errors isn't giving anything useful: they are infrequent in comparison and the
vast majority are obvious access errors like the namespace wasn't accessible.
Instead of trying to limit or change what we log, it seems prudent to just
remove it.

kubernetesListNamespaces isn't called nearly as much, but since it is the only
other list* tracking it seems like it should be removed to be consistent.

Signed-off-by: Tim deBoer <git@tdeboer.ca>