feat: extensions should stay disabled after restart
Disabled extensions should not start when you restart Podman Desktop.

Adds a new 'extensions.disabled' configuration property that stores an
array with the ids of disabled extensions. If an extension is disabled
it is read from disk but never activated.

To be able to only store/clear disabled extensions we need to know when
the user is intentionally enabling or disabling an extension (vs the
extension's normal start/stop with Podman Desktop). startExtension()
already exists but stopExtension() redirected to deactivateExtension(),
so I created a new stopExtension().

stopExtension() disables the extension, and both (manual) start,
removeExtensionPerUserRequest (for uninstall), and extension installer
(for install) make sure the extension is not disabled.

Fixes #268.

Signed-off-by: Tim deBoer <git@tdeboer.ca>