chore: more use of slide toggle
We have toggle switches in several places, but none of them were
the same code. This updates the existing SlideToggle to support a
few more cases (properties, aria, label on left or right) and adds
colors for light mode. The component is then reused for Settings >
Preferences and Settings > Proxies.

I had to make a couple judgement calls to make things consistent:
- The switch (circle) is bigger than before in Preferences.
- The proxy text gets dimmer when disabled/off, to match what was
  happening in preferences.
- I added missing hover states, making each one step brighter to
  match Buttons, except I made the off-hover state purple to match
  other controls to let you know you can click on it.

Signed-off-by: Tim deBoer <git@tdeboer.ca>