chore: table column ellipsis
Adds ellipsis to the standard table components when columns are too
narrow: column headers, simple columns, and duration columns. For
the header a div is added so that the sort indicators are always
visible, and DurationColumn can simplify by just being a SimpleColumn.

This does not cover any custom columns, of which there are many.

Fixes #7195.

Signed-off-by: Tim deBoer <git@tdeboer.ca>