chore: edge and light mode for modal dialog
Adds light mode colors for Modal dialog along with a defined 1px border so that
it is not washed out.

Initial light mode color values are taken from other mappings in the registry.

Also fixes a nit: the background should have the default cursor and not a
button cursor.

Fixes #7187.
First minor part of #7184.

Signed-off-by: Tim deBoer <git@tdeboer.ca>