chore: remove duplicate modal styling
When trying to test #7246 I noticed that there are several uses of Modal that
override the background color of Modal or try to add their own shadow or other
styling as per a very old design. This just removes this extraneous styling
so that these dialogs use what is in Modal (including light mode) and remaining
work can focus on what actually needs to change within the dialogs.

Should be more consistent or no change to existing modals, just clears out
unnecessary/duplicate code. IMHO SendFeedback is the only issue here since
two custom fields were being set to the (now) background color, not sure if
I should temporarily set them to another color or skip this for now.

Will need #7303 to see the rename image modal.

Part of #7184.

Signed-off-by: Tim deBoer <git@tdeboer.ca>