feat: table rows with children
Adds the ability for rows within a table to have children, very similar to
how container groups exist in the Containers view today.

Optional children function added to RowInformation to provide the
necessary API for consumers. Expand/collapse control matches what is in
the Containers view today.

In order to allow each row to be hover-able separately, a new parent div
creates the outer box for each parent/children group, and when children
exist these rows are added after the parent.

Since the UI package can't use stores directly like the Containers view
does, a new expanded array property has been exposed to allow consumers to
store/reload the list of expanded children. Since UI objects are recreated
each time, the list requires all expandable objects to have an id
property, and the list of ids is what is saved. This could be a
configurable property in the future, but everything we want to expand
today has an id.

Svelte sort animation conflicted with this change, so it has been removed.
If this PR is approved I'll open up an issue to track finding another way
to add it back.

Signed-off-by: Tim deBoer <git@tdeboer.ca>